feat: unified contents in tabs into one file
